refactor(component-container-refactor): moved IconTray to components (as isnt actually a container)
