+++ src/components/ChatboxResponse.js
@@ -1,0 +1,9 @@
+import React, { Component } from "react";
+
+class ChatBoxResponse extends Component {
+  render() {
+    return <p>ChatBoxResponse</p>;
+  }
+}
+
+export default ChatBoxResponse;
