+++ src/components/IconTray.js
